sort also paginated wearers
cleanup
